requirements.txt
langgraph 
langgraph-checkpoint-sqlite
langchain_community 
langchain-openai
langchain-google-genai 
langchain-groq 
langsmith
tavily-python
notion-client
slack_sdk
python-telegram-bot
twilio
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fastapi
uvicorn
python-multipart
python-dotenv
selenium
webdriver_manager
html2text
bs4
chromadb
unstructured
pdf2image
pytesseract
python-docx
pandas
pillow